break guard into two interfaces
