+++ src/Illuminate/Contracts/Auth/Guard.php
@@ -2,44 +2,8 @@
 
 namespace Illuminate\Contracts\Auth;
 
-interface Guard
+interface Guard extends StatelessGuard
 {
-    /**
-     * Determine if the current user is authenticated.
-     *
-     * @return bool
-     */
-    public function check();
-
-    /**
-     * Determine if the current user is a guest.
-     *
-     * @return bool
-     */
-    public function guest();
-
-    /**
-     * Get the currently authenticated user.
-     *
-     * @return \Illuminate\Contracts\Auth\Authenticatable|null
-     */
-    public function user();
-
-    /**
-     * Get the ID for the currently authenticated user.
-     *
-     * @return int|null
-     */
-    public function id();
-
-    /**
-     * Log a user into the application without sessions or cookies.
-     *
-     * @param  array  $credentials
-     * @return bool
-     */
-    public function once(array $credentials = []);
-
     /**
      * Attempt to authenticate a user using the given credentials.
      *
@@ -59,22 +23,6 @@ interface Guard
     public function basic($field = 'email');
 
     /**
-     * Perform a stateless HTTP Basic login attempt.
-     *
-     * @param  string  $field
-     * @return \Symfony\Component\HttpFoundation\Response|null
-     */
-    public function onceBasic($field = 'email');
-
-    /**
-     * Validate a user's credentials.
-     *
-     * @param  array  $credentials
-     * @return bool
-     */
-    public function validate(array $credentials = []);
-
-    /**
      * Log a user into the application.
      *
      * @param  \Illuminate\Contracts\Auth\Authenticatable  $user
@@ -82,14 +30,6 @@ interface Guard
      * @return void
      */
     public function login(Authenticatable $user, $remember = false);
-
-    /**
-     * Log the given user ID into the application without sessions or cookies.
-     *
-     * @param  mixed  $id
-     * @return bool
-     */
-    public function onceUsingId($id);
 
     /**
      * Log the given user ID into the application.
